DataViews: Add a context menu when right clicking on the table header
