+++ packages/dataviews/src/components/dataviews-view-config/index.tsx
@@ -14,20 +14,16 @@ import {
 	__experimentalToggleGroupControlOption as ToggleGroupControlOption,
 	__experimentalToggleGroupControlOptionIcon as ToggleGroupControlOptionIcon,
 	SelectControl,
-	__experimentalItemGroup as ItemGroup,
-	__experimentalItem as Item,
 	__experimentalGrid as Grid,
 	__experimentalVStack as VStack,
 	__experimentalHStack as HStack,
 	__experimentalHeading as Heading,
 	__experimentalText as Text,
 	privateApis as componentsPrivateApis,
-	BaseControl,
-	Icon,
 } from '@wordpress/components';
 import { __, _x } from '@wordpress/i18n';
 import { memo, useContext, useMemo } from '@wordpress/element';
-import { cog, check } from '@wordpress/icons';
+import { cog } from '@wordpress/icons';
 import warning from '@wordpress/warning';
 import { useInstanceId } from '@wordpress/compose';
 
@@ -36,9 +32,10 @@ import { useInstanceId } from '@wordpress/compose';
  */
 import { SORTING_DIRECTIONS, sortIcons, sortLabels } from '../../constants';
 import { VIEW_LAYOUTS } from '../../dataviews-layouts';
-import type { NormalizedField, View } from '../../types';
+import type { View } from '../../types';
 import DataViewsContext from '../dataviews-context';
 import InfiniteScrollToggle from './infinite-scroll-toggle';
+import { PropertiesSection } from './properties-section';
 import { unlock } from '../../lock-unlock';
 
 const { Menu } = unlock( componentsPrivateApis );
@@ -252,179 +249,6 @@ function ItemsPerPageControl() {
 	);
 }
 
-function FieldItem( {
-	field,
-	isVisible,
-	onToggleVisibility,
-}: {
-	field: NormalizedField< any >;
-	isVisible: boolean;
-	onToggleVisibility?: () => void;
-} ) {
-	return (
-		<Item onClick={ field.enableHiding ? onToggleVisibility : undefined }>
-			<HStack expanded justify="flex-start" alignment="center">
-				<div style={ { height: 24, width: 24 } }>
-					{ isVisible && <Icon icon={ check } /> }
-				</div>
-				<span>{ field.label }</span>
-			</HStack>
-		</Item>
-	);
-}
-
-function isDefined< T >( item: T | undefined ): item is T {
-	return !! item;
-}
-
-function FieldControl() {
-	const { view, fields, onChangeView } = useContext( DataViewsContext );
-
-	const togglableFields = [
-		view?.titleField,
-		view?.mediaField,
-		view?.descriptionField,
-	].filter( Boolean );
-
-	// Get all regular fields (non-locked) in their original order from fields prop
-	const regularFields = fields.filter(
-		( f ) =>
-			! togglableFields.includes( f.id ) &&
-			f.type !== 'media' &&
-			f.enableHiding !== false
-	);
-
-	if ( ! regularFields?.length ) {
-		return null;
-	}
-	const titleField = fields.find( ( f ) => f.id === view.titleField );
-	const previewField = fields.find( ( f ) => f.id === view.mediaField );
-	const descriptionField = fields.find(
-		( f ) => f.id === view.descriptionField
-	);
-
-	const lockedFields = [
-		{
-			field: titleField,
-			isVisibleFlag: 'showTitle',
-		},
-		{
-			field: previewField,
-			isVisibleFlag: 'showMedia',
-		},
-		{
-			field: descriptionField,
-			isVisibleFlag: 'showDescription',
-		},
-	].filter( ( { field } ) => isDefined( field ) );
-	const visibleFieldIds = view.fields ?? [];
-	const visibleRegularFieldsCount = regularFields.filter( ( f ) =>
-		visibleFieldIds.includes( f.id )
-	).length;
-
-	let visibleLockedFields = lockedFields.filter(
-		( { field, isVisibleFlag } ) =>
-			// @ts-expect-error
-			isDefined( field ) && ( view[ isVisibleFlag ] ?? true )
-	) as Array< {
-		field: NormalizedField< any >;
-		isVisibleFlag: string;
-	} >;
-
-	// If only one field (locked or regular) is visible, prevent it from being hidden
-	const totalVisibleFields =
-		visibleLockedFields.length + visibleRegularFieldsCount;
-	if ( totalVisibleFields === 1 ) {
-		if ( visibleLockedFields.length === 1 ) {
-			visibleLockedFields = visibleLockedFields.map( ( locked ) => ( {
-				...locked,
-				field: { ...locked.field, enableHiding: false },
-			} ) );
-		}
-	}
-
-	const hiddenLockedFields = lockedFields.filter(
-		( { field, isVisibleFlag } ) =>
-			// @ts-expect-error
-			isDefined( field ) && ! ( view[ isVisibleFlag ] ?? true )
-	) as Array< {
-		field: NormalizedField< any >;
-		isVisibleFlag: string;
-	} >;
-
-	return (
-		<VStack className="dataviews-field-control" spacing={ 0 }>
-			<BaseControl.VisualLabel>
-				{ __( 'Properties' ) }
-			</BaseControl.VisualLabel>
-			<VStack className="dataviews-view-config__properties" spacing={ 0 }>
-				<ItemGroup isBordered isSeparated size="medium">
-					{ visibleLockedFields.map( ( { field, isVisibleFlag } ) => {
-						return (
-							<FieldItem
-								key={ field.id }
-								field={ field }
-								isVisible
-								onToggleVisibility={ () => {
-									onChangeView( {
-										...view,
-										[ isVisibleFlag ]: false,
-									} );
-								} }
-							/>
-						);
-					} ) }
-
-					{ hiddenLockedFields.map( ( { field, isVisibleFlag } ) => {
-						return (
-							<FieldItem
-								key={ field.id }
-								field={ field }
-								isVisible={ false }
-								onToggleVisibility={ () => {
-									onChangeView( {
-										...view,
-										[ isVisibleFlag ]: true,
-									} );
-								} }
-							/>
-						);
-					} ) }
-
-					{ regularFields.map( ( field ) => {
-						// Check if this is the last visible field to prevent hiding
-						const isVisible = visibleFieldIds.includes( field.id );
-						const isLastVisible =
-							totalVisibleFields === 1 && isVisible;
-						const fieldToRender = isLastVisible
-							? { ...field, enableHiding: false }
-							: field;
-
-						return (
-							<FieldItem
-								key={ field.id }
-								field={ fieldToRender }
-								isVisible={ isVisible }
-								onToggleVisibility={ () => {
-									onChangeView( {
-										...view,
-										fields: isVisible
-											? visibleFieldIds.filter(
-													( fieldId ) =>
-														fieldId !== field.id
-											  )
-											: [ ...visibleFieldIds, field.id ],
-									} );
-								} }
-							/>
-						);
-					} ) }
-				</ItemGroup>
-			</VStack>
-		</VStack>
-	);
-}
-
 function SettingsSection( {
 	title,
 	description,
@@ -507,7 +331,7 @@ export function DataviewsViewConfigDropdown() {
 							) }
 							<InfiniteScrollToggle />
 							<ItemsPerPageControl />
-							<FieldControl />
+							<PropertiesSection />
 						</SettingsSection>
 					</VStack>
 				</DropdownContentWrapper>
