+++ packages/dataviews/src/components/dataviews-view-config/style.scss
@@ -58,3 +58,7 @@
 		}
 	}
 }
+
+.dataviews-view-config__label {
+	text-wrap: nowrap;
+}
